+++ nexus-writer/src/nexus/hdf5/mod.rs
@@ -1,0 +1,4 @@
+mod run_file;
+mod hdf5_writer;
+
+pub(crate) use run_file::RunFile;
